# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Voices of the Void - Void Dispatch API",
        "description": "API for Void Dispatch",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Host URL"
        }
    ],
    "paths": {
        "/api/effects": {
            "get": {
                "tags": [
                    "Effects"
                ],
                "summary": "Получить все внутриигровые эффекты",
                "operationId": "d08a997de614a65a8959b9545804367b",
                "parameters": [
                    {
                        "name": "recipient_token",
                        "in": "query",
                        "description": "Token игрока",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "player123"
                        }
                    },
                    {
                        "name": "current_minutes",
                        "in": "query",
                        "description": "Текущее внутриигровое время в минутах",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список эффектов игрока",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Effect"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "recipient_token не передан"
                    }
                }
            },
            "post": {
                "tags": [
                    "Effects"
                ],
                "summary": "Применить внутриигровой эффект",
                "operationId": "dc3582764cda6c71ab72f7c55a7ea248",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "recipient_token",
                                    "property",
                                    "value",
                                    "deliver_at_minutes"
                                ],
                                "properties": {
                                    "recipient_token": {
                                        "description": "Token игрока",
                                        "type": "string",
                                        "example": "player123"
                                    },
                                    "property": {
                                        "description": "Свойство, которое будет изменено",
                                        "type": "string",
                                        "enum": [
                                            "points",
                                            "food",
                                            "health",
                                            "sleep"
                                        ],
                                        "example": "points"
                                    },
                                    "value": {
                                        "description": "Значение (например +10, -5)",
                                        "type": "string",
                                        "example": "+50"
                                    },
                                    "deliver_at_minutes": {
                                        "description": "Внутриигровое время применения эффекта в минутах",
                                        "type": "integer",
                                        "example": 1440
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Effect created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Effect"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/effects/{effect}": {
            "delete": {
                "tags": [
                    "Effects"
                ],
                "summary": "Удалить внутриигровой эффект",
                "operationId": "9e87709a6f1de6621855c075b72ee39e",
                "parameters": [
                    {
                        "name": "effect",
                        "in": "path",
                        "description": "ID эффекта",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Effect deleted successfully"
                    },
                    "404": {
                        "description": "Effect not found"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Effects"
                ],
                "summary": "Обновить внутриигровой эффект",
                "operationId": "00a1f4c223884bab14e521f638b8f214",
                "parameters": [
                    {
                        "name": "effect",
                        "in": "path",
                        "description": "ID эффекта",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "recipient_token": {
                                        "description": "Token игрока",
                                        "type": "string",
                                        "example": "player123"
                                    },
                                    "property": {
                                        "description": "Свойство, которое будет изменено",
                                        "type": "string",
                                        "enum": [
                                            "points",
                                            "food",
                                            "health",
                                            "sleep"
                                        ],
                                        "example": "points"
                                    },
                                    "value": {
                                        "description": "Значение (например +10, -5)",
                                        "type": "string",
                                        "example": "+50"
                                    },
                                    "deliver_at_minutes": {
                                        "description": "Внутриигровое время применения эффекта в минутах",
                                        "type": "integer",
                                        "example": 1440
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Effect updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Effect"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Effect not found"
                    }
                }
            }
        },
        "/api/effects/consume": {
            "post": {
                "tags": [
                    "Effects"
                ],
                "summary": "Получить новые внутриигровые эффекты и пометить их примененными",
                "operationId": "ff195ddf019c04401cef60253315ac73",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "recipient_token",
                                    "current_minutes"
                                ],
                                "properties": {
                                    "recipient_token": {
                                        "description": "Токен игрока",
                                        "type": "string",
                                        "example": "player123"
                                    },
                                    "current_minutes": {
                                        "description": "Текущее игровое время в минутах",
                                        "type": "integer",
                                        "example": 1440
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Список новых внутриигровых эффектов, помеченных как примененные",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Effect"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/mails": {
            "get": {
                "tags": [
                    "Mails"
                ],
                "summary": "Получить все письма игрока",
                "operationId": "4fbe58b96147dd01f2c4ad51212293ca",
                "parameters": [
                    {
                        "name": "recipient_token",
                        "in": "query",
                        "description": "Token игрока",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "player123"
                        }
                    },
                    {
                        "name": "current_minutes",
                        "in": "query",
                        "description": "Текущее внутриигровое время в минутах",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список писем игрока",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Mail"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "recipient_token не передан"
                    }
                }
            },
            "post": {
                "tags": [
                    "Mails"
                ],
                "summary": "Добавить письмо",
                "operationId": "38fe45719214893161a95cce82b4cf7e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "recipient_token",
                                    "sender",
                                    "subject",
                                    "body",
                                    "deliver_at_minutes"
                                ],
                                "properties": {
                                    "recipient_token": {
                                        "description": "Token игрока, получателя письма",
                                        "type": "string",
                                        "example": "player123"
                                    },
                                    "sender": {
                                        "description": "Отправитель письма",
                                        "type": "string",
                                        "enum": [
                                            "Dr_Bao",
                                            "Prof_Lea",
                                            "Auto",
                                            "Dr_Max",
                                            "Dr_Ken",
                                            "Dr_Ena",
                                            "Dr_Ula",
                                            "Dr_Ler",
                                            "user",
                                            "Dr_Noa"
                                        ],
                                        "example": "Dr_Bao"
                                    },
                                    "subject": {
                                        "description": "Заголовок письма",
                                        "type": "string",
                                        "example": "Your mission"
                                    },
                                    "body": {
                                        "description": "Содержимое письма",
                                        "type": "string",
                                        "example": "Collect 10 samples"
                                    },
                                    "deliver_at_minutes": {
                                        "description": "Внутриигровое время доставки письма в минутах",
                                        "type": "integer",
                                        "example": 1440
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Mail created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Mail"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/mails/{mail}": {
            "delete": {
                "tags": [
                    "Mails"
                ],
                "summary": "Удалить письмо",
                "operationId": "84a8209660539635353c69f81b2c3933",
                "parameters": [
                    {
                        "name": "mail",
                        "in": "path",
                        "description": "ID письма",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Mail deleted successfully"
                    },
                    "404": {
                        "description": "Mail not found"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Mails"
                ],
                "summary": "Обновить письмо",
                "operationId": "138df9b7485b3ad1556708645f5aa3c3",
                "parameters": [
                    {
                        "name": "mail",
                        "in": "path",
                        "description": "ID письма",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "recipient_token": {
                                        "description": "Token игрока, получателя письма",
                                        "type": "string",
                                        "example": "player123"
                                    },
                                    "sender": {
                                        "description": "Отправитель письма",
                                        "type": "string",
                                        "enum": [
                                            "Dr_Bao",
                                            "Prof_Lea",
                                            "Auto",
                                            "Dr_Max",
                                            "Dr_Ken",
                                            "Dr_Ena",
                                            "Dr_Ula",
                                            "Dr_Ler",
                                            "user",
                                            "Dr_Noa"
                                        ],
                                        "example": "Dr_Bao"
                                    },
                                    "subject": {
                                        "description": "Заголовок письма",
                                        "type": "string",
                                        "example": "Your mission"
                                    },
                                    "body": {
                                        "description": "Содержимое письма",
                                        "type": "string",
                                        "example": "Collect 10 samples"
                                    },
                                    "deliver_at_minutes": {
                                        "description": "Внутриигровое время доставки письма в минутах",
                                        "type": "integer",
                                        "example": 1440
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Mail updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Mail"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Mail not found"
                    }
                }
            }
        },
        "/api/mails/consume": {
            "post": {
                "tags": [
                    "Mails"
                ],
                "summary": "Получить новые письма и пометить их прочитанными",
                "operationId": "464623f56995fe369627920ebfed3e14",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "recipient_token",
                                    "current_minutes"
                                ],
                                "properties": {
                                    "recipient_token": {
                                        "description": "Токен игрока, получателя письма",
                                        "type": "string",
                                        "example": "player123"
                                    },
                                    "current_minutes": {
                                        "description": "Текущее игровое время в минутах",
                                        "type": "integer",
                                        "example": 1440
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Список доставленных писем, помеченных как прочитанные",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Mail"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Effect": {
                "properties": {
                    "id": {
                        "description": "Class Effect",
                        "type": "integer"
                    },
                    "recipient_token": {
                        "type": "string"
                    },
                    "property": {
                        "type": "string"
                    },
                    "value": {
                        "type": "string",
                        "nullable": true
                    },
                    "deliver_at_minutes": {
                        "type": "integer"
                    },
                    "is_applied": {
                        "type": "boolean"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Mail": {
                "properties": {
                    "id": {
                        "description": "Class Mail",
                        "type": "integer"
                    },
                    "recipient_token": {
                        "type": "string"
                    },
                    "sender": {
                        "type": "string"
                    },
                    "subject": {
                        "type": "string"
                    },
                    "body": {
                        "type": "string"
                    },
                    "deliver_at_minutes": {
                        "type": "integer"
                    },
                    "is_read": {
                        "type": "boolean"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Effects",
            "description": "Effects"
        },
        {
            "name": "Mails",
            "description": "Mails"
        }
    ]
}